refactor: 파일 구조 변경
- main branch와의 동기화를 위한 디렉터리 구조 수정
