+++ linkup/src/main/java/com/learningcrew/linkup/exception/ErrorResponse.java
@@ -1,4 +1,4 @@
-package com.learningcrew.linkup.common.exception;
+package com.learningcrew.linkup.exception;
 
 import lombok.Builder;
 import lombok.Getter;
