+++ linkup/src/main/java/com/learningcrew/linkup/linker/command/domain/service/UserDomainService.java
@@ -1,8 +1,8 @@
 package com.learningcrew.linkup.linker.command.domain.service;
 
 import com.learningcrew.linkup.common.constants.StatusType;
-import com.learningcrew.linkup.common.exception.BusinessException;
-import com.learningcrew.linkup.common.exception.ErrorCode;
+import com.learningcrew.linkup.exception.BusinessException;
+import com.learningcrew.linkup.exception.ErrorCode;
 import com.learningcrew.linkup.linker.command.domain.aggregate.Role;
 import com.learningcrew.linkup.linker.command.domain.aggregate.Status;
 import com.learningcrew.linkup.linker.command.domain.aggregate.User;
