+++ linkup/src/main/java/com/learningcrew/linkup/meeting/command/application/service/ParticipantReviewCommandService.java
@@ -1,7 +1,7 @@
 package com.learningcrew.linkup.meeting.command.application.service;
 
-import com.learningcrew.linkup.common.exception.BusinessException;
-import com.learningcrew.linkup.common.exception.ErrorCode;
+import com.learningcrew.linkup.exception.BusinessException;
+import com.learningcrew.linkup.exception.ErrorCode;
 import com.learningcrew.linkup.meeting.command.application.dto.request.ParticipantReviewCreateRequest;
 import com.learningcrew.linkup.meeting.command.domain.aggregate.ParticipantReview;
 import com.learningcrew.linkup.meeting.command.domain.repository.ParticipantReviewRepository;
