--- linkup/src/main/java/com/learningcrew/linkup/meeting/query/dto/response/InterestedMeetingResponse.java
@@ -1,10 +1,0 @@
-package com.learningcrew.linkup.meeting.query.dto.response;
-
-//import lombok.Builder;
-//import lombok.Getter;
-//
-//@Getter
-//@Builder
-//public class InterestedMeetingResponse {
-//    private final InterestedMeetingDTO interestedMeeting;
-//}
