+++ linkup/src/main/java/com/learningcrew/linkup/security/handler/RestAuthenticationEntryPoint.java
@@ -1,9 +1,9 @@
 package com.learningcrew.linkup.security.handler;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.learningcrew.linkup.common.exception.CustomJwtException;
-import com.learningcrew.linkup.common.exception.ErrorCode;
-import com.learningcrew.linkup.common.exception.ErrorResponse;
+import com.learningcrew.linkup.exception.security.CustomJwtException;
+import com.learningcrew.linkup.exception.ErrorCode;
+import com.learningcrew.linkup.exception.ErrorResponse;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
