+++ linkup/src/main/java/com/learningcrew/linkup/security/jwt/JwtTokenProvider.java
@@ -1,13 +1,12 @@
 package com.learningcrew.linkup.security.jwt;
 
-import com.learningcrew.linkup.common.exception.CustomJwtException;
-import com.learningcrew.linkup.common.exception.ErrorCode;
+import com.learningcrew.linkup.exception.security.CustomJwtException;
+import com.learningcrew.linkup.exception.ErrorCode;
 import io.jsonwebtoken.*;
 import io.jsonwebtoken.io.Decoders;
 import io.jsonwebtoken.security.Keys;
 import jakarta.annotation.PostConstruct;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.stereotype.Component;
 
 import javax.crypto.SecretKey;
